+++ frontend/src/components/AuthPage/Auth.css
@@ -36,7 +36,7 @@ form.login-form__group {
 .input__text {
   height: 3rem;
   width: 100%;
-  box-shadow: rgba(0, 0, 0, 0.02) 0px 1px 3px 0px, rgba(27, 31, 35, 0.15) 0px 0px 0px 1px;
+  box-shadow: var(--input-box-shadow);
   border-radius: 0.2em;
   padding-left: 1em;
 }
