+++ frontend/src/components/Dashboard/assets/Dashboard.css
@@ -6,6 +6,7 @@
   --border-bottom-color: orange;
   --custom-orange: #e48f10;
   --box-shadow: rgba(0, 0, 0, 0.059) 0px 4px 12px;
+  --input-box-shadow: rgba(0, 0, 0, 0.02) 0px 1px 3px 0px, rgba(27, 31, 35, 0.15) 0px 0px 0px 1px;
 }
 
 /* --------- Mobile --------- */
