+++ frontend/src/components/Events/InvitationForm/assets/InvitationForm.css
@@ -1,0 +1,25 @@
+/* ---- .invitation-form-content ---- */
+
+.invitation-form-content {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-items: center;
+  align-content: center;
+  padding: 2em;
+  gap: 30px;
+}
+
+/* ---- planning, filters ---- */
+
+.planning,
+.filters {
+  display: flex;
+  flex-direction: column;
+  min-width: fit-content;
+}
+
+.planning > *,
+.filters > * {
+  margin-bottom: 1em;
+}
