+++ frontend/src/components/Profile/assets/Profile.css
@@ -156,7 +156,8 @@ h2 {
   display: flex;
   flex-direction: row;
   align-items: center;
-  padding: 0.5em;
+  justify-content: space-between;
+  padding: 2.5em;
   border-top: 0.2em solid var(--dashboard-background);
   background-color: var(--profile-section-background-color);
 }
