+++ frontend/src/components/Shared/components/SearchPopup/SearchPopup.css
@@ -71,6 +71,7 @@ input.group.name {
   border-radius: 0.3em;
   background-color: white;
   padding: 0.2em 1em;
+  box-shadow: var(--input-box-shadow);
 }
 
 /* -- Disable default styles -- */
@@ -137,7 +138,7 @@ input {
 .inputs {
   display: flex;
   flex-direction: row;
-  box-shadow: var(--box-shadow);
+  box-shadow: var(--input-box-shadow);
   height: 2.5em;
   font-size: 0.9em;
   border-radius: 0.3em;
